er i glade?
